requirements.txt
Django==2.2.6
Django-ckeditor==5.7.1